camera fixes and others
Created a room template for you to duplicate and then create levels with
Fixed purple background controls positioning.
Fixed camera movement
when pressing the left and right buttons
So the camera doesn’t follow when in the middle of the screen
So the camera follows if you are off screen
The camera moves with you vertically only if you are on the ground
At the start of the game, the camera will move with you vertically even when you are in the air. However, it will not follow you vertically in the air once you have landed
Fixed positioning of cat when using melee attack on the ground
Extended the left barrier to 320. I added a left barrier object so you can see where the barrier is. This is just so the player can’t see the grass end.
Fixed some layering.